Moved logic to search component, made sure nothing is displayed when searchvalue is empty
