+++ src/pages/dashboard/index.js
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import SearchIcon from '../../assets/icons/searchIcon';
 import Button from '../../components/button';
 import Card from '../../components/card';
@@ -6,113 +6,17 @@ import CreatePostModal from '../../components/createPostModal';
 import TextInput from '../../components/form/textInput';
 import Posts from '../../components/posts';
 import useModal from '../../hooks/useModal';
-import SearchResults from './search';
+import StudentSearchResults from './search';
+import useAuth from '../../hooks/useAuth';
 import './style.css';
 
 const Dashboard = () => {
-  const profiles = [
-    {
-      id: 1,
-      userId: 1,
-      firstName: 'John',
-      lastName: 'Doe',
-      specialism: 'Full-stack developer',
-      githubUrl: 'https://github.com/johndoe'
-    },
-    {
-      id: 2,
-      userId: 2,
-      firstName: 'Jane',
-      lastName: 'Smith',
-      specialism: 'Frontend engineer',
-      githubUrl: 'https://github.com/janesmith'
-    },
-    {
-      id: 3,
-      userId: 3,
-      firstName: 'Michael',
-      lastName: 'Johnson',
-      specialism: 'Backend developer',
-      githubUrl: 'https://github.com/michaeljohnson'
-    },
-    {
-      id: 4,
-      userId: 4,
-      firstName: 'Emily',
-      lastName: 'Davis',
-      specialism: 'DevOps engineer',
-      githubUrl: 'https://github.com/emilydavis'
-    },
-    {
-      id: 5,
-      userId: 5,
-      firstName: 'Chris',
-      lastName: 'Brown',
-      specialism: 'Data scientist',
-      githubUrl: 'https://github.com/chrisbrown'
-    },
-    {
-      id: 6,
-      userId: 6,
-      firstName: 'Sophia',
-      lastName: 'Wilson',
-      specialism: 'UI/UX designer',
-      githubUrl: 'https://github.com/sophiawilson'
-    },
-    {
-      id: 7,
-      userId: 7,
-      firstName: 'Daniel',
-      lastName: 'Martinez',
-      specialism: 'Software engineer',
-      githubUrl: 'https://github.com/danielmartinez'
-    },
-    {
-      id: 8,
-      userId: 8,
-      firstName: 'Olivia',
-      lastName: 'Garcia',
-      specialism: 'Product manager',
-      githubUrl: 'https://github.com/oliviagarcia'
-    },
-    {
-      id: 9,
-      userId: 9,
-      firstName: 'Liam',
-      lastName: 'Taylor',
-      specialism: 'Security analyst',
-      githubUrl: 'https://github.com/liamtaylor'
-    },
-    {
-      id: 10,
-      userId: 10,
-      firstName: 'Emma',
-      lastName: 'Andersson',
-      specialism: 'Backend developer',
-      githubUrl: 'https://github.com/emmaanderson'
-    }
-  ];
-
   const [searchVal, setSearchVal] = useState('');
-  const [searchUsers, setSearchUsers] = useState([]);
-
-  useEffect(() => {
-    setSearchUsers(profiles); // Save all users for filtering
-  }, []);
+  const { role } = useAuth();
 
   const onChange = (e) => {
     const value = e.target.value;
     setSearchVal(value);
-
-    // Filter users based on search value using firstName and lastName
-    if (value) {
-      const filteredUsers = profiles.filter((user) =>
-        `${user.firstName} ${user.lastName}`.toLowerCase().includes(value.toLowerCase())
-      );
-      setSearchUsers(filteredUsers);
-    } else {
-      setSearchUsers([]); // Clear search results if input is empty
-    }
   };
 
   const onSubmit = (e) => {
@@ -154,9 +58,11 @@ const Dashboard = () => {
             />
           </form>
         </Card>
-
-        <SearchResults searchUsers={searchUsers} />
-
+        {role === 'teacher' ? (
+          <p className="teacher-view-message">Teacher-specific view goes here</p>
+        ) : (
+          <StudentSearchResults searchVal={searchVal} />
+        )}
         <Card>
           <h4>My Cohort</h4>
         </Card>
